move stores out of directories
