--- src/store/modal/modal-store.ts
@@ -1,15 +1,0 @@
-import {IModalStore} from '../store-types'
-import { action } from 'easy-peasy';
-
-export const ModalStore: IModalStore = {
-    openedCreateProjectModal: false,
-    openCreateProjectModal: action((state) => {
-        state.openedCreateProjectModal = true;
-    }),
-    closeCreateProjectModal: action((state) => {
-        state.openedCreateProjectModal = false;
-    }),
-}
-
-// export default new AuthStore()
-export default ModalStore
